refactor(engineer_worktree): extract claim helpers to claim.rs (#1266)
Promote engineer_worktree/mod.rs (778 LOC) by extracting PID + starttime
sentinel helpers into a sibling claim.rs module:

- read_pid_starttime, format_engineer_claim, is_pid_alive
- EngineerClaim struct + read_engineer_claim/_full + claim_is_live

mod.rs: 778 -> 650 (still partial; needs sweep_orphaned_worktrees split next)
claim.rs: 138 LOC

All 17 engineer_worktree tests pass.
